function array args were actually optional args
Woops! I saw "[int" and assumed it was an array type
specifications (like in jni), NOPE! it was actually "[int]" in the man
pages and it meant it was an optional argument. So I
have gone and expanded those functions into versions with and without
the arg.

In glsl-spec "[type" still means array of the type